refactor: extract board_run service from route handler (#96 step 2)
